build: 4.22.1
feat： 新增lamp-sa-token-ext模块，重新实现 sa-token 的SaRequestForServlet

fix: 修复跨服务调用接口时，传递的请求头Token无法被sa-token解析的问题

refactor: uri鉴权时，查询系统所有的URI接口缓存至redis

refactor： lamp-cloud 和 lamp-boot 项目合并为一个工程，实现1个工程同时支持单体版和微服务版

refactor： 原来的api层变更为facade层，以方便适配单体版或微服务版

facade层包含3层：lamp-xx-api、lamp-xx-boot-impl、lamp-xx-cloud-impl
api层用于定义接口，boot-impl层用于单体版实现api层接口，cloud-impl层用于微服务版实现api层接口
对于微服务版：a服务调用b服务接口时，在lamp-a-biz层加入lamp-b-api层的依赖，并在lamp-a-server层加入lamp-b-cloud-impl依赖即可
对于单体版：a服务(lamp-a-biz)调用b服务(lamp-b-biz)接口时，在lamp-a-biz层加入lamp-b-api层的依赖，并在lamp-boot-server层加入lamp-b-boot-impl依赖即可
refactor(lamp-generator)：代码生成器的模版文件抽取到lamp-generator-biz层

refactor(lamp-generator)：优化代码生成器服务的依赖关系

refactor(lamp-oauth)：lamp-oauth-userinfo模块和lamp-oauth-userinfo模块更名为lamp-login-user-facade

lamp-login-user-resolver：负责拦截@LoginUser注解，并定义接口
lamp-login-user-boot-impl：负责单体版实现具体的查询逻辑
lamp-login-user-cloud-impl：负责微服务版实现具体的查询逻辑
refactor(lamp-oauth): 删除没使用的接口

refactor(doc.yml): swagger公共配置类代码和配置文件优化

refactor(lamp-common-api): 删除lamp-common-api模块

+++ lamp-base/lamp-base-biz/src/main/java/top/tangyh/lamp/base/service/system/impl/BaseRoleServiceImpl.java
@@ -13,6 +13,7 @@ import top.tangyh.basic.base.service.impl.SuperCacheServiceImpl;
 import top.tangyh.basic.context.ContextUtil;
 import top.tangyh.basic.database.mybatis.conditions.Wraps;
 import top.tangyh.basic.database.mybatis.conditions.query.LbQueryWrap;
+import top.tangyh.basic.exception.BizException;
 import top.tangyh.basic.model.cache.CacheKey;
 import top.tangyh.basic.utils.ArgumentAssert;
 import top.tangyh.basic.utils.CollHelper;
@@ -95,6 +96,11 @@ public class BaseRoleServiceImpl extends SuperCacheServiceImpl<BaseRoleManager, 
     public boolean removeByIds(Collection<Long> idList) {
         if (idList.isEmpty()) {
             return true;
+        }
+        List<BaseRole> baseRoles = superManager.listByIds(idList);
+        long count = baseRoles.stream().filter(item -> item.getReadonly() != null && item.getReadonly()).count();
+        if (count > 0) {
+            throw new BizException("内置角色不允许删除");
         }
         // 员工的角色
         baseEmployeeRoleRelManager.deleteByRole(idList);
